requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
python-dotenv
google-generativeai